requirements.txt
annotated-types==0.6.0
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
evdev==1.7.0
ffmpeg-python==0.2.0
filelock==3.13.3
fsspec==2024.3.1
future==1.0.0
huggingface-hub==0.22.2
idna==3.6
importlib_metadata==7.1.0
Jinja2==3.1.3
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
openai-whisper==20231117
packaging==24.0
paho-mqtt==2.0.0
pocketsphinx==5.0.3
PyAudio==0.2.14
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pynput==1.7.6
python-xlib==0.33
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
safetensors==0.4.2
six==1.16.0
sounddevice==0.4.6
SpeechRecognition==3.10.1
sympy==1.12
tdqm==0.0.1
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
transformers==4.39.2
triton==2.2.0
typing_extensions==4.10.0
urllib3==2.2.1
whisper_mic==1.4.2
zipp==3.18.1
